Fix balloon print when there's a failure
Don't mark a balloon as printed unless it successfully completes the print
job. There could still be a failure printing and it may look odd in the UI
when 'auto print' is enabled but it doesn't show as printed yet, but this
should generally improve things in case of problems like print queue failure
or printer unplugged.